+++ lib/td_df_web/controllers/template_controller.ex
@@ -2,6 +2,8 @@ defmodule TdDfWeb.TemplateController do
   use TdDfWeb, :controller
   use PhoenixSwagger
 
+  import Canada, only: [can?: 2]
+
   alias TdDf.Templates
   alias TdDf.Templates.Preprocessor
   alias TdDf.Templates.Template
@@ -38,7 +40,10 @@ defmodule TdDfWeb.TemplateController do
   end
 
   def create(conn, %{"template" => template}) do
-    with {:ok, %Template{} = template} <- Templates.create_template(template) do
+    claims = conn.assigns[:current_resource]
+
+    with {:can, true} <- {:can, can?(claims, manage(template))},
+         {:ok, %Template{} = template} <- Templates.create_template(template) do
       conn
       |> put_status(:created)
       |> put_resp_header("location", Routes.template_path(conn, :show, template))
@@ -86,10 +91,11 @@ defmodule TdDfWeb.TemplateController do
   end
 
   def update(conn, %{"id" => id, "template" => template_params}) do
+    claims = conn.assigns[:current_resource]
     template = Templates.get_template!(id)
-    update_params = Map.drop(template_params, [:name])
 
-    with {:ok, %Template{} = template} <- Templates.update_template(template, update_params) do
+    with {:can, true} <- {:can, can?(claims, manage(template))},
+         {:ok, %Template{} = template} <- Templates.update_template(template, template_params) do
       render(conn, "show.json", template: template)
     end
   end
@@ -107,9 +113,11 @@ defmodule TdDfWeb.TemplateController do
   end
 
   def delete(conn, %{"id" => id}) do
+    claims = conn.assigns[:current_resource]
     template = Templates.get_template!(id)
 
-    with {:ok, %Template{}} <- Templates.delete_template(template) do
+    with {:can, true} <- {:can, can?(claims, manage(template))},
+         {:ok, %Template{}} <- Templates.delete_template(template) do
       send_resp(conn, :no_content, "")
     end
   end