+++ priv/repo/migrations/20190118134400_remove_template_relation.exs
@@ -1,7 +1,0 @@
-defmodule TdDf.Repo.Migrations.RemoveTemplateRelation do
-  use Ecto.Migration
-
-  def change do
-    drop(table(:template_relations))
-  end
-end